The program is now functionally complete.
Adds custom argparse Action subclass for validating the identifier
pattern program option

Adds a method open_enc() that opens a file using either the built-in
open() or the open method found in the codecs module for opening UTF-16
encoded files. This is now incorporated throughout the program.

Adds gather_sample_ids() as an alternative to the
gather_hierarchical_sample_ids(), to be used when all the sequence data
is contained within a single folder

The write_mapping_file() method uses a file handle instead of opening a
file handle from an input file name. The program now uses the argparse
module's ability to directly open a file during option parsing.

The scrobble_fasta_dir() method now works for both hierarchical and
single-directory layouts, and handles UTF-16 input

The argparse parser now handles a --version flag to print out the
program version, and is more user-friendly with the identifier pattern
option. Finally it has options for instructing the program to look for
quality files in addition to fasta files, and has a flag for handling
UTF-16 input files

Main now handles both hierarchy and single directory layouts.