Adds script filter_keep_otus_by_sample.py
This script allows the opposite of the QIIME script filter_otus_by_sample.py.
Instead of listing the Sample IDs that will be removed from the new file, this
script allows users to list which Sample IDs will be kept in the new file; often
a smaller list than the negative.