Adds script biom_phyla_summary.py
Takes in a BIOM abundance table and prints out a summary of the percent
abundance of the top 20 Families, Genera, and Species.

Also currently lists the percent abundance of important species in the oral
cavity (color complexes). This will be made an option in a later release.
Along with a more generic ability to pass in a list of any phylogenetic level
or specific OTU.