Add filter_biom.py script
Filters BIOM format file to based on mapping file provided.
Dual filtering of input file, first on “sample” and then on
“observation” axis is performed.
Entries in “observation” axis whose abundance sum to 0 are removed, and
have an option to be saved in a separate file using “-fo” parameter.
Script is compatible with Python2 and Python3 syntax, and conforms to
PEP8 guidelines.