Fixes example instructions for the PCoA script.
Instead of relying only on beta_diversity_through_plots.py, the instructions are now more general. The (now) 3 steps proceed from beta_diversity.py to principal_coordinates.py to PCoA.py. The original instructions are preseved in an alternate to steps 1 and 2.

Also updates the help text for the --color_by option to make clear all the options for grouping categories.