Adds new TSV input option to core_overlap_plot
Instead of limiting users to supplying results files from QIIME's compute_core_microbiome script, they can now give a file containing a simple list of names for each group. Those names will be used to determine the overlap for the plot.

This opens up the script to be used, for example, with metagenomic functional data or any other kind of data.

If users do provide a multi-column TSV file, only the first column will be used.