Allow users to keep OTU IDs in the iTol output
Instead of always attempting to translate the OTU IDs to a short name based on the taxonomy, this commit adds a new parameter: keep_otuids.

This may be useful for non-standard taxonomy or for datasets using NCBI taxonomy IDs since the iTol webapp can itself look up their information and replace IDs with taxonomic names.