Fixes sample ids being interpreted as numbers
When loading the distance matrix, if the sample IDs are just numbers, Pandas will make the indices of the DataFrame integers instead of strings. There's no simple way (that I've found) to force just the index entries to be strings when loading a table. So we'll just use python to convert to str in the two places that access the index.