Modifies output format
Removes the column headers in the output file as they aren't really necessary, and without them users can simply copy and paste the output columns into another file.

The program also now writes a blank entry for IDs that are not found. This has a side benefit of dealing nicely with input files that may or may not have headers.